Update int32 to int64

+++ SolanaClassLibrary/BusinessLogic/GetFeesProcessor.cs
@@ -24,10 +24,10 @@ namespace SolanaClassLibrary.BusinessLogic
             gfmr.result.value.feeCalculator = new GetFeesModel.Feecalculator();
 
             gfmr.jsonrpc = job["jsonrpc"].ToString();
-            gfmr.result.context.slot = Convert.ToInt32(job["result"]["context"]["slot"]);
+            gfmr.result.context.slot = Convert.ToInt64(job["result"]["context"]["slot"]);
             gfmr.result.value.blockhash = job["result"]["value"]["blockhash"].ToString();
             gfmr.result.value.feeCalculator.lamportsPerSignature = Convert.ToInt32(job["result"]["value"]["feeCalculator"]["lamportsPerSignature"]);
-            gfmr.result.value.lastValidSlot = Convert.ToInt32(job["result"]["value"]["lastValidSlot"]);
+            gfmr.result.value.lastValidSlot = Convert.ToInt64(job["result"]["value"]["lastValidSlot"]);
             gfmr.id = Convert.ToInt32(job["id"]);
 
             return gfmr;
